dan fixed navigation to login
