+++ src/components/button.tsx
@@ -3,6 +3,7 @@ import { cn } from "../utils/styling";
 type Props = {
   children: React.ReactNode;
   variant?: "primary" | "secondary";
+  type?: "button" | "submit";
   size?: "default" | "small";
   className?: string;
   onClick?: () => void;
@@ -12,12 +13,14 @@ export function Button({
   children,
   variant = "primary",
   size = "default",
+  type = "button",
   className,
   onClick,
 }: Props) {
   return (
     <button
       onClick={onClick}
+      type={type}
       className={cn(
         "rounded-lg font-semibold w-full",
         variant === "primary" ? "bg-yellow disabled:bg-yellow/50" : null,
