+++ src/layouts/navbar-layout.tsx
@@ -1,4 +1,4 @@
-import { Outlet, useNavigate } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import { NavBar } from "../components/nav-bar";
 import {
   FilmIcon,
@@ -6,19 +6,12 @@ import {
   QueueListIcon,
   UserIcon,
 } from "@heroicons/react/24/solid";
-import { useEffect } from "react";
+import { useAuthenticated } from "../hooks/useAuthenticated";
 
 export function NavBarLayout() {
   // navigate to login page if no authentifictaion
-  const navigate = useNavigate();
 
-  useEffect(() => {
-    let authState = JSON.parse(localStorage.getItem("auth-state") || "{}");
-    if (!authState || !authState.authState) {
-      authState = false;
-      navigate("/login");
-    }
-  });
+  useAuthenticated();
 
   const navItems = [
     { icon: HomeIcon, route: "/" },
